fix(claude): read account-prefixed Desktop token caches
